+++ lib/yokogaki.dart
@@ -4,9 +4,11 @@ library;
 
 // Widgets
 export 'src/widgets/horizontal_text.dart';
+export 'src/widgets/horizontal_rich_text.dart';
 
 // Models
 export 'src/models/horizontal_text_style.dart';
+export 'src/models/horizontal_text_span.dart';
 export 'src/models/ruby_text.dart';
 export 'src/models/kenten.dart';
 export 'src/models/warichu.dart';